fix(cli): split text busy retry by platform
