+++ internal/pipeline/textfilebusy_unix.go
@@ -1,0 +1,12 @@
+//go:build !windows
+
+package pipeline
+
+import (
+	"errors"
+	"syscall"
+)
+
+func isTextFileBusy(err error) bool {
+	return errors.Is(err, syscall.ETXTBSY)
+}
