+++ internal/pipeline/textfilebusy_windows.go
@@ -1,0 +1,7 @@
+//go:build windows
+
+package pipeline
+
+func isTextFileBusy(error) bool {
+	return false
+}
